# DATAMONGO-2046 - Performance optimizations.