Various optimizations to improve CPU and memory consumption.

**Baseline**

```
Benchmark                                                                     Mode  Cnt        Score        Error  Units
MappingMongoConverterBenchmark.readObjectWith2Properties                     thrpt   10  1595760,841 ± 190172,611  ops/s
MappingMongoConverterBenchmark.readObjectWith2PropertiesAnd1NestedObject     thrpt   10   745325,414 ±  31819,541  ops/s
MappingMongoConverterBenchmark.readObjectWith2PropertiesConstructorCreation  thrpt   10  1545948,636 ±  71914,291  ops/s
MappingMongoConverterBenchmark.readObjectWithListAndMapsOfComplexType        thrpt   10   114421,125 ±   2570,836  ops/s
MappingMongoConverterBenchmark.writeObjectWith2PropertiesAnd1NestedObject    thrpt   10   990447,535 ±  18082,689  ops/s
MappingMongoConverterBenchmark.writeObjectWithListAndMapsOfComplexType       thrpt   10   150329,330 ±   1552,529  ops/s
```

**After optimizations** (without `DATACMNS-1366`)

```
Benchmark                                                                     Mode  Cnt        Score        Error  Units
MappingMongoConverterBenchmark.readObjectWith2Properties                     thrpt   10  1901418,878 ±  42860,397  ops/s
MappingMongoConverterBenchmark.readObjectWith2PropertiesAnd1NestedObject     thrpt   10   878169,600 ± 140822,915  ops/s
MappingMongoConverterBenchmark.readObjectWith2PropertiesConstructorCreation  thrpt   10  1691408,752 ±  29058,524  ops/s
MappingMongoConverterBenchmark.readObjectWithListAndMapsOfComplexType        thrpt   10   139606,761 ±   2933,151  ops/s
MappingMongoConverterBenchmark.writeObjectWith2PropertiesAnd1NestedObject    thrpt   10  1002344,779 ±  19669,362  ops/s
MappingMongoConverterBenchmark.writeObjectWithListAndMapsOfComplexType       thrpt   10   158590,534 ±   2457,356  ops/s
```

**After optimizations** (with `DATACMNS-1366`)

```
Benchmark                                                                     Mode  Cnt        Score       Error  Units
MappingMongoConverterBenchmark.readObjectWith2Properties                     thrpt   10  1952981,262 ± 29559,740  ops/s
MappingMongoConverterBenchmark.readObjectWith2PropertiesAnd1NestedObject     thrpt   10   964554,056 ±  7183,559  ops/s
MappingMongoConverterBenchmark.readObjectWith2PropertiesConstructorCreation  thrpt   10  1704016,207 ± 35992,424  ops/s
MappingMongoConverterBenchmark.readObjectWithListAndMapsOfComplexType        thrpt   10   148036,200 ±  1835,007  ops/s
MappingMongoConverterBenchmark.writeObjectWith2PropertiesAnd1NestedObject    thrpt   10  1048958,420 ±  8158,864  ops/s
MappingMongoConverterBenchmark.writeObjectWithListAndMapsOfComplexType       thrpt   10   160894,020 ±   952,022  ops/s
```

---

Related tickets: [DATACMNS-1366](https://jira.spring.io/browse/DATACMNS-1366), [DATACMNS-1366](https://jira.spring.io/browse/DATAMONGO-2046).